Test and fix that closing a tied handle may die
This required the change that File::Replace::SingleHandle::CLOSE needs
to always be fatal, so that its behavior is consistent. Previously, a
close failure on whichever handle was closed first would just return a
false value, while closing the second handle would trigger the call to
"finish", where a close failure would be fatal.

This commit also fixes a spurious warning on autocancel.